adds pages
Added Home and Navbar
